Moved service related files into service
